backend-project/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.4
django-allauth==0.62.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
factory-boy==3.3.0
Faker==25.2.0
filelock==3.14.0
fsspec==2024.3.1
fugashi==1.3.2
googlemaps==4.10.0
huggingface-hub==0.23.0
idna==3.7
ipadic==1.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.5.10
requests==2.31.0
safetensors==0.4.3
setuptools==69.2.0
six==1.16.0
sqlparse==0.5.0
sympy==1.12
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.40.2
typing_extensions==4.11.0
urllib3==2.2.1
wheel==0.43.0
